+++ src/constants/index.ts
@@ -1,2 +1,1 @@
-export { default as SAMPLE } from './sample';
-export { default as MEDIA_TYPE } from './mediaType';
+export * from './mediaType';
