+++ src/constants/mediaType.ts
@@ -1,6 +1,4 @@
-const MEDIA_TYPE = {
+export const MEDIA_TYPE = {
   MOVIE: 1,
   TV: 2,
 };
-
-export default MEDIA_TYPE;
