--- src/constants/sample.ts
@@ -1,5 +1,0 @@
-const SAMPLE = {
-  name: 'sample',
-};
-
-export default SAMPLE;
