+++ src/hooks/index.ts
@@ -1,1 +1,1 @@
-export { default as useSearch } from './useSearch';
+export * from './useSearch';
